Fix useLocalStorage
The old hook was buggy, and wouldn't recheck if the input key changed.
So when it started as `undefined` for the first render or two, it would
search and find it undefined, and never find the true value under the
slug. So, always read the value from local storage. This causes some UI
flashing after submitting new lines, but works otherwise. If I have more
time tomorrow, I will try to come up with something better.